+++ engine/src/platform/windows/windows_time.cpp
@@ -1,11 +1,9 @@
 // Copyright (c) 2020 udv. All rights reserved.
 
 #include "real/platform/windows/windows_time.hpp"
-#include "real/api/gl/gl_headers.hpp"
 
 namespace real::platform {
-	float time() {
-		// TODO: use native time
+	double time() {
 		return glfwGetTime();
 	}
 }
